style: 메인페이지 반응형 스타일 적용
